# src/window-renderer/src/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"WindowMaterialType": {
				"type": "string",
				"enum": [
					"CAS",
					"F1A",
					"F2A",
					"FIX",
					"FLD",
					"FLS",
					"FRO",
					"PF1A",
					"PF2A",
					"PRT1A",
					"PRT2A",
					"SIL",
					"SLA",
					"SLF",
					"SRAF",
					"SRLA",
					"VAS"
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "src",
		"version": "1.4.0",
		"description": "REST microservice to generate scaled raster images of window types based on dimensions and configuration",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/windows/drawWindow": {
			"get": {
				"operationId": "DrawWindowImage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Generates a raster image of a window based on height, width, and material type\r\nand returns it as a PNG stream response.\r\n\r\nThe image is scaled proportionally to match the input dimensions\r\nand reflects the visual representation of the specified window material type:\r\n\r\n- 'CAS': Cassonetto\r\n- 'F1A': Finestra 1 Anta\r\n- 'F2A': Finestra 2 Ante\r\n- 'FIX': Fisso centrale\r\n- 'FLD': Fisso laterale destro\r\n- 'FLS': Fisso laterale sinistro\r\n- 'FRO': Frontalino\r\n- 'PF1A': Portafinestra 1 Anta\r\n- 'PF2A': Portafinestra 2 Ante\r\n- 'PRT1A': Portoncino 1 Anta\r\n- 'PRT2A': Portoncino 2 Ante\r\n- 'SIL': Scorrevole in linea\r\n- 'SLA': Sopraluce apribile\r\n- 'SLF': Sopraluce fisso\r\n- 'SRAF': Scorrevole ribalta con anta fissa\r\n- 'SRLA': Scorrevole ribalta con laterale apribile\r\n- 'VAS': Vasistas",
				"summary": "Generate and download window raster image as PNG",
				"tags": [
					"Windows"
				],
				"security": [],
				"parameters": [
					{
						"description": "Height of the window in millimeters",
						"in": "query",
						"name": "height",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Width of the window in millimeters",
						"in": "query",
						"name": "width",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Material type of the window",
						"in": "query",
						"name": "materialType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/WindowMaterialType"
						}
					},
					{
						"description": "Optional: include wire cover rendering",
						"in": "query",
						"name": "wireCover",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Optional: direction of glass opening ('OT_DX' or 'OT_SX')",
						"in": "query",
						"name": "glassType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"OT_DX",
								"OT_SX"
							]
						}
					},
					{
						"description": "Optional: glass type ('GT_TRASPARENTE' or 'GT_OPACO')",
						"in": "query",
						"name": "openingType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"GT_TRASPARENTE",
								"GT_OPACO"
							]
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	],
	"tags": [
		{
			"name": "Windows",
			"description": "Windows Renderer"
		}
	]
}